Convert IrcServer to Typescript
