+++ src/bridge/IrcBridge.js
@@ -9,7 +9,7 @@ var MatrixHandler = require("./MatrixHandler.js");
 var MemberListSyncer = require("./MemberListSyncer.js");
 var IdentGenerator = require("../irc/IdentGenerator.js");
 var Ipv6Generator = require("../irc/Ipv6Generator.js");
-var IrcServer = require("../irc/IrcServer.js");
+const { IrcServer } = require("../irc/IrcServer.js");
 var ClientPool = require("../irc/ClientPool");
 var IrcEventBroker = require("../irc/IrcEventBroker");
 var BridgedClient = require("../irc/BridgedClient");
